Rename library to Insets
