+++ insets/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=accompanist-insets
+POM_NAME=Accompanist Insets library
+POM_PACKAGING=aar
