+++ insets/src/androidTest/AndroidManifest.xml
@@ -14,5 +14,9 @@
   ~ limitations under the License.
   -->
 
-<manifest package="dev.chrisbanes.accompanist.insetter">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="dev.chrisbanes.accompanist.insets.test">
+
+    <uses-sdk android:targetSdkVersion="30" />
+
 </manifest>
