--- insetter/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=accompanist-insetter
-POM_NAME=Accompanist Insetter library
-POM_PACKAGING=aar
