+++ settings.gradle
@@ -30,5 +30,5 @@ include ':picasso'
 include ':glide'
 include ':imageloading-core'
 include ':imageloading-testutils'
-include ':insetter'
+include ':insets'
 include ':sample'
